refactor pytorch distribution tests
